chore/refactor-productareacomponent - Refactor the large ProductAreaComponent and ProductAreaSteps classes into smaller classes.
